third_party/blink/web_tests/flag-specific/highdpi/fast/shapes/shape-outside-floats/shape-outside-floats-image-margin-002-expected.txt

This test requires the Ahem font. The green content should wrap around the blue rounded rectangle.

X

X
X

X
XXXX
PASS successfullyParsed is true

TEST COMPLETE

FAIL imageShapeRect('a').top should be 0. Was -0.6666641235351562.
FAIL imageShapeRect('b').top should be 50. Was 48.66667175292969.
FAIL imageShapeRect('b').right should be 300. Was 299.9999885559082.
FAIL imageShapeRect('c').top should be 75. Was 73.33333587646484.
FAIL imageShapeRect('c').right should be 300. Was 299.9999885559082.
FAIL imageShapeRect('d').top should be 125. Was 122.66667175292969.
FAIL imageShapeRect('d').right should be within 1 of 292. Was 293.3333320617676.
FAIL imageShapeRect('e').top should be 150. Was 147.33334350585938.
FAIL imageShapeRect('e').right should be 0. Was 261.3333282470703.
